add comment for Docusaurus v4 refactor helmet
